added VUHF-RX added DensityMap and eye plot improved file source callback when device was closed many updates

+++ EZ-USB/RX-FFT/LibRXFFT/Components/DirectX/Drawables/Docks/TestDock.cs
@@ -10,8 +10,8 @@ namespace LibRXFFT.Components.DirectX.Drawables.Docks
     public class TestDock : Dock
     {
         public override string Title { get; set; }
-        public override int Height { get; set; }
-        public override int Width { get; set; }
+        public override float Height { get; set; }
+        public override float Width { get; set; }
         public override bool Sticky { get; set; }
         public override bool HideOthers { get; set; }
